Moved AwwIncentiveAggregationHelper to own file
