+++ custom/icds_reports/models/aggregate.py
@@ -10,9 +10,6 @@ from custom.icds_reports.const import (AGG_CCS_RECORD_BP_TABLE,
     AGG_CHILD_HEALTH_PNC_TABLE, AGG_CHILD_HEALTH_THR_TABLE,
     AGG_COMP_FEEDING_TABLE, AGG_DAILY_FEEDING_TABLE,
     AGG_GROWTH_MONITORING_TABLE, AGG_INFRASTRUCTURE_TABLE, AWW_INCENTIVE_TABLE)
-from custom.icds_reports.utils.aggregation import (
-    AwwIncentiveAggregationHelper
-)
 from dateutil.relativedelta import relativedelta
 from django.db import connections, models, transaction
 from six.moves import range
@@ -20,6 +17,7 @@ from six.moves import range
 from custom.icds_reports.utils.aggregation_helpers.agg_ccs_record import AggCcsRecordAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.agg_child_health import AggChildHealthAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.awc_infrastructure import AwcInfrastructureAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.aww_inactive import AwwIncentiveAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.birth_preparedness_forms import \
     BirthPreparednessFormsAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.ccs_record_monthly import CcsRecordMonthlyAggregationHelper
